% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R, R/environment_grid.r
\name{add_agents}
\alias{add_agents}
\alias{add_agents.tidyabm_env}
\alias{add_agents.tidyabm_env_grid}
\title{Add agents to the environment}
\usage{
add_agents(.tidyabm, agent, n, ...)

\method{add_agents}{tidyabm_env}(.tidyabm, agent, n, ...)

\method{add_agents}{tidyabm_env_grid}(.tidyabm, agent, n, initial_position = "random", ...)
}
\arguments{
\item{.tidyabm}{the \link{tidyabm_env} object to which agents should be added}

\item{agent}{an \link{tidyabm_agent} object (the blueprint) to replicate from}

\item{n}{number of agents in the style of \code{agent} to add}

\item{...}{other arguments to pass to sub functions}

\item{initial_position}{one of the following:
\itemize{
\item \code{random} independently choose a random position per agent
\item \code{ordered-rowwise} fill up the grid from the top left in the order that
the agents have been added, filling up row by row
\item \code{ordered-colwise} fill up the grid from the top left in the order that
the agents have been added, filling up column by column
\item \code{connect-four} independently choose a random column for each agent to be
added to the grid, filling up from the bottom (just like the board game)
\item a vector of length 2 with an \code{x} and a \code{y} position to place the
particular agent at (note that this only works if one agent is added at
a time)
\item a list of length \code{n} with a vector of length 2 at each spot indicating
\code{x} and \code{y} of the particular position of this particular agent
}}
}
\value{
\link{tidyabm} object
}
\description{
Create an agent first (through \link{create_agent}). This agent can be seen as
a blueprint. Then, use this function here to add as many replicates of this
blueprint agent as necessary. They will be added independently so that they
do not interfere with each other throughout runtime.
}
\examples{
a1 <- create_agent() \%>\%
  set_characteristic(age = 'young')
a2 <- create_agent() \%>\%
  set_characteristic(age = 'old')
m <- create_grid_environment(seed = 1024, size = 50) \%>\%
  add_agents(a1, 100) \%>\%
  add_agents(a2, 100)

}
