% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{add_variable}
\alias{add_variable}
\alias{add_variable.tidyabm}
\title{Add a flexible variable}
\usage{
add_variable(.tidyabm, ..., .overwrite = FALSE)

\method{add_variable}{tidyabm}(.tidyabm, ..., .overwrite = FALSE)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm} object}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Name-value pairs.
The name gives the name of the variable. To avoid misunderstandings use
unique variable names, also between agents and environments. Try to also
avoid starting your name with a point/dot (\code{.}) because some internally
added names use that pattern. Good variable names also do not have any
whitespace and are all lowercase; use underscores to separate any words.
It is recommended to also avoid the use of minus signs. Hence, a good
variable name looks like \code{this_is_a_great_name}.
The value has to be one of:
\itemize{
\item a single value (e.g., a character, numeric, logical, vector ...)
\item a function to be run at every iteration of the modelling process
}

Keep in mind that for single values that are not expected to change,
\link{set_characteristic} is a better-fitting method as it is not re-run at
every iteration of the modelling process.

For functions you may use a function which gets called at every iteration
with two arguments, \code{me} (which is the \link{tidyabm} object at that specific
point in time) and \code{abm} (which is the whole \link{tidyabm} object). Note that
for \link{tidyabm_env} objects these two arguments are the same but for
\link{tidyabm_agent} objects they are different in that the first (\code{me}) is
the current agent and the second (\code{abm}) is the whole environment model.
If you write your own functions or need to provide additional arguments to
functions, use the style of anonymized functions directly in-line (through
\code{function(me, abm) ...} or \verb{\\(me, abm) ...}).}

\item{.overwrite}{if \code{FALSE} (the default), variables with the same name
will not be overwritten (a warning will be issued)}
}
\value{
a \link{tidyabm} object
}
\description{
A variable is expected to updated at various (if not all) iterations. It
gets initiated at the start with the status quo of the model and is
accessible throughout runtime. Compared to a characteristic, it thus
requires runtime resources but is flexibly updated at every iteration.
}
\examples{
create_agent() \%>\%
  set_characteristic(age = 15) \%>\%
  add_variable(share_same_age = function(me) {
                neighbors <- get_neighbors(me)
                return(sum(neighbors$age == me$age)/nrow(neighbors))
              }) \%>\%
  add_variable(feels_welcome = \(me) share_same_age >= .50)

create_grid_environment(seed = 9896, size = 5) \%>\%
  add_variable(m_age = \(me) mean(me$agents$age, na.rm = T))

}
