% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{convert_agents_to_tibble}
\alias{convert_agents_to_tibble}
\title{Convert the list of agent \code{tidyabm_agent} objects to a \link{tibble}}
\usage{
convert_agents_to_tibble(.tidyabm)
}
\arguments{
\item{.tidyabm}{the \code{tidyabm_env} object holding some agents}
}
\value{
a \link{tibble}
}
\description{
Returns a tibble with all agents in the same order as the agent objects
were initially added. Columns represent a unique identifier named \code{.id},
all agent characteristics, and all agent variables. Not that, if not at
least one tick has passed, agent variables are \code{NA}. Any characteristics
that some agents have while others do not will be \code{NA} for those agents
who do not share this particular characteristic. Typically,
characteristics/variables added internally through a particular
environment are prefixed with \code{.}.
}
\examples{
create_grid_environment(seed = 4583, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  convert_agents_to_tibble()

}
