% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{convert_network_to_tidygraph}
\alias{convert_network_to_tidygraph}
\title{Convert current environment to \link{tidygraph} object}
\usage{
convert_network_to_tidygraph(.tidyabm)
}
\arguments{
\item{.tidyabm}{the current \code{tidyabm_env_network} object}
}
\value{
a \code{tbl_graph} object
}
\description{
Convert current environment to \link{tidygraph} object
}
\examples{
a <- create_agent() \%>\%
  add_rule('connect to a random other agent',
           .consequence = \(me, abm) {
             me \%>\%
               network_connect(get_random_agent(abm, me)) \%>\%
               return()
           })
create_network_environment(seed = 183468,
                           is_directed = TRUE) \%>\%
  add_agents(a,
             n = 5) \%>\%
  init() \%>\%
  tick() \%>\%
  convert_network_to_tidygraph()

}
