% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{create_agent}
\alias{create_agent}
\title{Create an agent}
\usage{
create_agent()
}
\value{
a \link{tidyabm} object of subclass \code{tidyabm_agent}
}
\description{
This function is used to create agent blueprints that have to, in a
second step, be added to an environment through \link{add_agents}.
}
\examples{
create_agent()

}
