% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{create_grid_environment}
\alias{create_grid_environment}
\title{Create an orthogonal grid environment}
\usage{
create_grid_environment(seed, size = NULL, x = NULL, y = NULL)
}
\arguments{
\item{seed}{a single number (integer) to initiate random-number generation}

\item{size}{number to indicate side length for square}

\item{x}{numeric side length of side x (omitted if \code{size} is set)}

\item{y}{numeric side length of side y (omitted if \code{size} is set)}
}
\value{
\code{tidyabm} object
}
\description{
Creates an environment with a specified two-dimensional grid. Either define
the grid as a square by setting one side's \code{size} \emph{or} define the grid as a
rectangle by setting both \code{x} and \code{y}.

The seed is used to explicitly document random-number generation
initiation in your script.
}
\examples{
create_grid_environment(seed = 42, size = 25)

create_grid_environment(seed = 123, x = 25, y = 5)

}
