% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{create_network_environment}
\alias{create_network_environment}
\title{Create a network environment}
\usage{
create_network_environment(seed, is_directed = FALSE)
}
\arguments{
\item{seed}{a single number (integer) to initiate random-number generation}

\item{is_directed}{if true, connections/links in the network have a direction
(i.e., they are arrows); default is FALSE}
}
\value{
\code{tidyabm} object
}
\description{
Creates an environment without any space limitations. However, agents
therein can connect with each other and form a network.

The seed is used to explicitly document random-number generation
initiation in your script.
}
\examples{
create_network_environment(seed = 43)

create_network_environment(seed = 124, is_directed = TRUE)

}
