% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_characteristic}
\alias{get_characteristic}
\title{Get an environment's characteristic value}
\usage{
get_characteristic(me, characteristic)
}
\arguments{
\item{me}{the \code{tidyabm_env} environment object}

\item{characteristic}{character string naming the characteristic}
}
\value{
the characteristic value (or NULL if not found)
}
\description{
Get an environment's characteristic value
}
\seealso{
Other utilities: 
\code{\link{get_random_agent}()},
\code{\link{get_variable}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{network_connect}()},
\code{\link{network_get_neighbors}()},
\code{\link{network_spread}()},
\code{\link{remove_agent}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
