% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_random_agent}
\alias{get_random_agent}
\title{Get a random agent from all agents in the environment}
\usage{
get_random_agent(abm, me = NULL)
}
\arguments{
\item{abm}{the whole environment model (\code{abm})}

\item{me}{if provided as an agent then the returned (random) agent will not
be this one (i.e., \code{me} is omitted)}
}
\value{
a \link{tidyabm_agent} object
}
\description{
Get a random agent from all agents in the environment
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_variable}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
