% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_variable}
\alias{get_variable}
\title{Get an environment's variable value}
\usage{
get_variable(me, variable)
}
\arguments{
\item{me}{the \code{tidyabm_env} environment object}

\item{variable}{character string naming the variable}
}
\value{
the variable value (or NULL if variable is not found)
}
\description{
Get an environment's variable value
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_random_agent}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{network_connect}()},
\code{\link{network_spread}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
