% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{grid_get_free_neighboring_spots}
\alias{grid_get_free_neighboring_spots}
\title{Get a \link{tibble} with two columns (.x and .y) of spots that are free}
\usage{
grid_get_free_neighboring_spots(agent, abm, which = "o")
}
\arguments{
\item{agent}{the agent for whom the neighbors should be collected (\code{me})}

\item{abm}{the whole environment model}

\item{which}{one of the following:
\itemize{
\item \code{o} to get up to 8 neighboring spots surrounding \code{me} (default)
\item \code{+} to get up to 4 neighboring spots directly next to \code{me} (no diagonals)
\item \code{-} to get up to 2 x-axis neighboring spots, one on each side of \code{me}
\item \code{|} to get up to 2 y-axis neighboring spots, one above and one below \code{me}
\item \verb{--} to get up to 4 x-axis neighboring spots, two on each side of \code{me}
(this actually works up to \verb{-----})
\item \code{||} to get up to 4 y-axis neighboring spots, two above and two below
\code{me} (this actually works up to \verb{|||||})
}}
}
\value{
a \link{tibble} of neighboring spots (with a \code{.x} and a \code{.y} column) in
no particular order that are free
}
\description{
Get a \link{tibble} with two columns (.x and .y) of spots that are free
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_random_agent}()},
\code{\link{get_variable}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
