% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{grid_get_neighbors}
\alias{grid_get_neighbors}
\title{Get a \link{tibble} of direct neighbors}
\usage{
grid_get_neighbors(agent, abm, which = "o")
}
\arguments{
\item{agent}{the agent for whom the neighbors should be collected (\code{me})}

\item{abm}{the whole environment model}

\item{which}{one of the following:
\itemize{
\item \code{o} to get up to 8 neighbors surrounding \code{me} (default)
\item \code{+} to get up to 4 neighbors directly next to \code{me} (no diagonals)
\item \code{-} to get up to 2 x-axis neighbors, one on each side of \code{me}
\item \code{|} to get up to 2 y-axis neighbors, one above and one below \code{me}
\item \verb{--} to get up to 4 x-axis neighbors, two on each side of \code{me}
(this actually works up to \verb{-----})
\item \code{||} to get up to 4 y-axis neighbors, two above and two below \code{me}
(this actually works up to \verb{|||||})
}}
}
\value{
a \link{tibble} of neighboring agents in no particular order with their
characteristics and variables set
}
\description{
Get a \link{tibble} of direct neighbors
}
\seealso{
Other utilities: 
\code{\link{get_random_agent}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_move}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
