% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{grid_move}
\alias{grid_move}
\title{Move an agent to the proposed new x/y position (if it is empty)}
\usage{
grid_move(agent, abm, new_x, new_y)
}
\arguments{
\item{agent}{the agent for whom the neighbors should be collected (\code{me})}

\item{abm}{the whole environment model (\code{abm})}

\item{new_x}{numeric new x position}

\item{new_y}{numeric new y position}
}
\value{
a \code{tidyabm_agent} object
}
\description{
If the new position is not empty or does not exist, the agent will not
move (obviously) and a warning will be issued.
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_random_agent}()},
\code{\link{get_variable}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
