% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{init}
\alias{init}
\alias{init.tidyabm_env}
\title{Initialize the environment and prepare everything for take-off}
\usage{
init(.tidyabm)

\method{init}{tidyabm_env}(.tidyabm)
}
\arguments{
\item{.tidyabm}{the \link{tidyabm_env} object}
}
\value{
a \link{tidyabm_env} object
}
\description{
Initialize the environment and prepare everything for take-off
}
\examples{
create_grid_environment(seed = 46444, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init()

}
