% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{is_directed}
\alias{is_directed}
\title{Check if given \code{tidyabm_env_network} is directed}
\usage{
is_directed(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if given \code{tidyabm_env_network} is directed
}
