% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{is_ended}
\alias{is_ended}
\title{Check if a given environment has ended}
\usage{
is_ended(.tidyabm)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm_env} object}
}
\value{
logical
}
\description{
Check if a given environment has ended
}
