% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{is_tickable}
\alias{is_tickable}
\title{Check if a given environment is ready to \link{tick}/\link{iterate}}
\usage{
is_tickable(.tidyabm)
}
\arguments{
\item{.tidyabm}{a \code{tidyabm_env} object}
}
\value{
logical
}
\description{
Check if a given environment is ready to \link{tick}/\link{iterate}
}
