% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{is_tidyabm}
\alias{is_tidyabm}
\title{Check if a provided object is of type \code{tidyabm}}
\usage{
is_tidyabm(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if a provided object is of type \code{tidyabm}
}
