% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{is_tidyabm_agent}
\alias{is_tidyabm_agent}
\title{Check if a provided object is of type \code{tidyabm_agent}}
\usage{
is_tidyabm_agent(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if a provided object is of type \code{tidyabm_agent}
}
