% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{is_tidyabm_env}
\alias{is_tidyabm_env}
\title{Check if a provided object is of type \code{tidyabm_env}}
\usage{
is_tidyabm_env(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if a provided object is of type \code{tidyabm_env}
}
