% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{is_tidyabm_env_grid}
\alias{is_tidyabm_env_grid}
\title{Check if a provided object is of type \code{tidyabm_env_grid}}
\usage{
is_tidyabm_env_grid(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if a provided object is of type \code{tidyabm_env_grid}
}
\keyword{internal}
