% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{is_tidyabm_env_network}
\alias{is_tidyabm_env_network}
\title{Check if a provided object is of type \code{tidyabm_env_network}}
\usage{
is_tidyabm_env_network(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if a provided object is of type \code{tidyabm_env_network}
}
\keyword{internal}
