% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{iterate}
\alias{iterate}
\alias{iterate.tidyabm_env}
\title{Take off and iterate}
\usage{
iterate(.tidyabm, max_iterations = 50, verbose = TRUE, visualize = FALSE, ...)

\method{iterate}{tidyabm_env}(.tidyabm, max_iterations = 50, verbose = TRUE, visualize = FALSE, ...)
}
\arguments{
\item{.tidyabm}{the \link{tidyabm_env} object}

\item{max_iterations}{number of iterations after which iterations stop (if
it was not stopped earlier by other means of convergence, e.g. through
environment rules).}

\item{verbose}{if TRUE (the default), a status message is printed at the end}

\item{visualize}{if TRUE, each tick ends with a visualization if this has
been implemented for the respective environment; default is FALSE}

\item{...}{further arguments to pass on to visualize}
}
\value{
a \link{tidyabm_env} object
}
\description{
This is the main function to run through your model (i.e., the
environment with all the agents in it). This function is just a loop
that calls \link{tick} again and again until a stop signal is sent or until
the \code{max_iterations} limit has been reached.

\strong{CAUTION}: This may take a while!

A freshly \link{init}'ed environment starts with the first iteration, of course.
However, you can also continue iterating a model. The model remembers the
tick it stopped and proceeds there the next time (see examples).

For more information on how an individual \link{tick} is structured, see the
documentation there.
}
\examples{
create_grid_environment(seed = 4583, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init() \%>\%
  iterate()

m <- create_grid_environment(seed = 321456, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init() \%>\%
  iterate(max_iterations = 10)

# continue with the 11th iteration and iterate all the way through to #15
m \%>\%
  iterate(max_iterations = 5)

}
\seealso{
\link{tick}
}
