% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_grid.r
\name{matrix_get_neighbors}
\alias{matrix_get_neighbors}
\title{Get neighbors for a given position (but inside the matrix)}
\usage{
matrix_get_neighbors(.tidyabm, x, y)
}
\arguments{
\item{.tidyabm}{the \code{tidyabm_env_grid} object}

\item{x}{numeric x of the middle}

\item{y}{numeric y of the middle}
}
\value{
a named length-8 vector with either -1 (field does not exist), 0
(field is empty) or the index (!) of the corresponding agent (to address
them directly via \code{attr(.tidyabm, 'agents')[[index]]})
}
\description{
Get neighbors for a given position (but inside the matrix)
}
