% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{n_agents}
\alias{n_agents}
\title{Total number of agents in a model}
\usage{
n_agents(x, ...)
}
\arguments{
\item{x}{An \code{abm_model} or \code{abm_result} object.}

\item{...}{Unused.}
}
\value{
An integer.
}
\description{
Total number of agents in a model
}
