% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{network_connect}
\alias{network_connect}
\title{Add a link between source and target agents}
\usage{
network_connect(source_agent, target_agent)
}
\arguments{
\item{source_agent}{the agent from whom the link should start}

\item{target_agent}{the agent where the link should end}
}
\value{
a \code{tidyabm_agent} object
}
\description{
If the network is directed, then the arrow points from source to target.
If the network is undirected, the link does not contain any arrow.
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_random_agent}()},
\code{\link{get_variable}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
