% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{network_get_neighbors}
\alias{network_get_neighbors}
\title{Get a \link{tibble} of direct neighbors}
\usage{
network_get_neighbors(agent, abm, only_use_outdegree = is_directed(abm))
}
\arguments{
\item{agent}{the agent for whom the neighbors should be collected (\code{me})}

\item{abm}{the whole environment model}

\item{only_use_outdegree}{if true (default in directed models) then
neighbors are only considered neighbors if \code{agent} points at them, not
vice versa (in undirected models, this parameter is ignored)}
}
\value{
a \link{tibble} of neighboring agents in no particular order with their
characteristics and variables set
}
\description{
Get a \link{tibble} of direct neighbors
}
\seealso{
Other utilities: 
\code{\link{get_characteristic}()},
\code{\link{get_random_agent}()},
\code{\link{get_variable}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_free_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()},
\code{\link{network_connect}()},
\code{\link{network_spread}()},
\code{\link{remove_agent}()},
\code{\link{stop_abm}()}
}
\concept{utilities}
