% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R, R/tidyabm_class.R
\name{new_tidyabm_env}
\alias{new_tidyabm_env}
\alias{new_tidyabm}
\title{Output constructor for \code{tidyabm_env} objects}
\usage{
new_tidyabm_env(data, particular_suffix, class_params = list())

new_tidyabm(
  data,
  class_suffix = "",
  particular_sub_class = NULL,
  class_params = list()
)
}
\arguments{
\item{data}{a \link{tibble}}

\item{particular_suffix}{string to append to main class \code{tidyabm_env_}}

\item{class_params}{list with parameters initially relevant to store}

\item{class_suffix}{string to suffix sub class name with; has to be one of
\verb{ } (empty), \code{env}, or \code{agent}.}

\item{particular_sub_class}{if set to character then this is the first class
name to use, even before the suffixed sub class name (\code{NULL} meaning no
particular sub class is the default)}
}
\value{
a \link{tidyabm} object with optional subclass of \code{particular_sub_class}
and subclass of \verb{tidyabm_*suffix*}
}
\description{
Output constructor for \code{tidyabm_env} objects

General \code{tidyabm} output constructor
}
