% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{odd}
\alias{odd}
\alias{odd.tidyabm_env}
\title{Extract compact information for an ODD protocol}
\usage{
odd(.tidyabm, ...)

\method{odd}{tidyabm_env}(.tidyabm)
}
\arguments{
\item{.tidyabm}{the \code{tidyabm_env} object}

\item{...}{other arguments passed to particular types of environment}
}
\value{
a \link{tibble} of length 7 (the ODD categories)
}
\description{
ODD (Overview, Design concepts and Details) protocols are a suggested
standard for describing agent-based models (Grimm et al. 2006, 2010, 2020).
As Grimm et al. (2020) state, "ODD model descriptions \link{...} are based on
written text and intended to be read by humans." This function provides all
the details the environment and agent specifications contain that might be
relevant for the ODD to be written. It returns the 2020 ODD framework with
all the collected information pre-filled into the respective category.
}
\examples{
create_grid_environment(seed = 4583, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init() \%>\%
  iterate() \%>\%
  odd()

}
\references{
Grimm, V., U. Berger, F. Bastiansen, S. Eliassen, V. Ginot, J.
Giske, J. Goss-Custard, T. Grand, S. Heinz, G. Huse, A. Huth, J. U. Jepsen,
C. Jørgensen, W. M. Mooij, B. Müller, G. Pe’er, C. Piou, S. F. Railsback,
A. M. Robbins, M. M. Robbins, E. Rossmanith, N. Rüger, E. Strand, S.
Souissi, R. A. Stillman, R. Vabø, U. Visser, & D. L. DeAngelis (2006). A
standard protocol for describing individual-based and agent-based models.
Ecological Modelling, 198, 115-296. 10.1016/j.ecolmodel.2006.04.023

Grimm, V., U. Berger, D. L. DeAngelis, J. G. Polhill, J. Giske, & S. F.
Railsback (2010). The ODD protocol: A review and first update. Ecological
Modelling 221, 2760-2768. 10.1016/j.ecolmodel.2010.08.019.

Grimm, V., S. F. Railsback, C. E. Vincenot, U. Berger, C. Gallagher, D. L.
DeAngelis, B. Edmonds, J. Ge, J. Giske, J. Groeneveld, A. S. A. Johnston,
A. Milles, J. Nabe-Nielsen, G. Polhill, V. Radchuk, M.-S. Rohwäder, R. A.
Stillman, J. C. Thiele, & D. Ayllón (2020). The ODD Protocol for describing
agent-based and other simulation models: A second update to improve
clarity, replication, and structural realism. Journal of Artificial
Societies and Social Simulation, 23(2), 7. 10.18564/jasss.4259
}
