% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{if_else}
\alias{case_when}
\alias{between}
\alias{coalesce}
\alias{n}
\alias{tibble}
\alias{.data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:between]{between()}}, \code{\link[dplyr:case_when]{case_when()}}, \code{\link[dplyr:coalesce]{coalesce()}}, \code{\link[dplyr:if_else]{if_else()}}, \code{\link[dplyr:n]{n()}}}

  \item{rlang}{\code{\link[rlang:.data]{.data}}}

  \item{tibble}{\code{\link[tibble:tibble]{tibble()}}}
}}

