% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{remove_rule}
\alias{remove_rule}
\alias{remove_rule.tidyabm}
\title{Remove a certain rule}
\usage{
remove_rule(.tidyabm, label)

\method{remove_rule}{tidyabm}(.tidyabm, label)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm} object}

\item{label}{string used to describe the rule}
}
\value{
a \link{tidyabm} object
}
\description{
Remove a certain rule
}
\examples{
create_agent() \%>\%
  set_characteristic(age = 15) \%>\%
  add_rule('check minority', age >= 18, stop_abm) \%>\%
  remove_rule('check minority')

create_grid_environment(seed = 13122, size = 5) \%>\%
  add_rule('no more agents', 'todo') \%>\%
  remove_rule('no more agents', 'todo')

}
