% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{retain_new_data_in_prior_object}
\alias{retain_new_data_in_prior_object}
\alias{retain_new_data_in_prior_object.tidyabm}
\title{Only update data in an already existent \link{tidyabm} object}
\usage{
retain_new_data_in_prior_object(prior_object, new_data)

\method{retain_new_data_in_prior_object}{tidyabm}(prior_object, new_data)
}
\arguments{
\item{prior_object}{a \link{tidyabm} object}

\item{new_data}{a \link{tibble}}
}
\value{
a \link{tidyabm} object, just like \code{prior_object}
}
\description{
Only update data in an already existent \link{tidyabm} object
}
