% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{rule_mirror_links}
\alias{rule_mirror_links}
\title{Fill up .indegree reflected on those newly set .outdegree links}
\usage{
rule_mirror_links(me, abm)
}
\arguments{
\item{me}{the \code{tidyabm_env_network} object pushed from the .mirror_links rule}

\item{abm}{unused}
}
\value{
the updated \code{tidyabm_env_network} object
}
\description{
Fill up .indegree reflected on those newly set .outdegree links
}
