% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_network.R
\name{rule_spread}
\alias{rule_spread}
\title{Distribute information as requested by \code{network_spread}}
\usage{
rule_spread(me, abm)
}
\arguments{
\item{me}{the \code{tidyabm_env_network} object pushed from the .spread rule}

\item{abm}{unused}
}
\value{
the updated \code{tidyabm_env_network} object
}
\description{
Distribute information as requested by \code{network_spread}
}
