% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{set_characteristic}
\alias{set_characteristic}
\alias{set_characteristic.tidyabm}
\title{Set an initial characteristic}
\usage{
set_characteristic(
  .tidyabm,
  ...,
  .overwrite = FALSE,
  .suppress_warnings = FALSE
)

\method{set_characteristic}{tidyabm}(
  .tidyabm,
  ...,
  .overwrite = FALSE,
  .suppress_warnings = FALSE
)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm} object}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> name-value pairs.
The name lets you identify the characteristic. To avoid misunderstandings
use unique names, also between agents and environments. Try to also
avoid starting your name with a point/dot (\code{.}) because some internally
added names use that pattern. Good names also do not have any whitespace
and are all lowercase; use underscores to separate any words. It is
recommended to also avoid the use of minus signs. Hence, a good
characteristic name looks like \code{this_is_a_great_name}.
The value is treated as a single value (e.g., a character, numeric,
logical, vector ...).}

\item{.overwrite}{if \code{FALSE} (the default), characteristics with the same
name will not be overwritten (a warning will be issued)}

\item{.suppress_warnings}{if \code{TRUE}, \code{.overwrite} will not yield any
warnings; default is FALSE, though}
}
\value{
a \link{tidyabm} object
}
\description{
A characteristic is appended to the respective object right from the start.
It is accessible throughout runtime and can also change if a rule chooses
to do so. However, compared to a variable, it is not calculated/determined
in each iteration.

If you need to roll out a characteristic's distribution across a range of
agents (e.g., a characteristic that's supposed to be normally distributed
across a group or all of the agents), see
\link{distribute_characteristic_across_agents}.
}
\examples{
create_grid_environment(seed = 3, size = 5) \%>\%
  set_characteristic(condition = 'small')

create_agent() \%>\%
  set_characteristic(age = 15)

create_agent() \%>\%
  set_characteristic(age = 15,
                     gender = 'm')

create_agent() \%>\%
  set_characteristic(embed_vector = c(23.2, 12.4, 4.2, -3.2, 13.2, -0.1))

}
