% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{stop_abm}
\alias{stop_abm}
\title{Claim a simulation to end}
\usage{
stop_abm(abm, me = NULL)
}
\arguments{
\item{abm}{the whole environment model (\code{abm})}

\item{me}{unnecessary but provided here for ease of use}
}
\value{
a \link{tidyabm_env} object
}
\description{
Claim a simulation to end
}
\seealso{
Other utilities: 
\code{\link{get_random_agent}()},
\code{\link{grid_get_free_neighboring_spots}()},
\code{\link{grid_get_neighbors}()},
\code{\link{grid_move}()}
}
\concept{utilities}
