% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R, R/environment.R, R/tidyabm_class.R
\name{update_values.tidyabm_agent}
\alias{update_values.tidyabm_agent}
\alias{update_values.tidyabm_env}
\alias{update_values}
\title{Update an actual value from a characteristic or a variable}
\usage{
\method{update_values}{tidyabm_agent}(.tidyabm, values, is_characteristic = FALSE)

\method{update_values}{tidyabm_env}(.tidyabm, values, is_characteristic = FALSE)

update_values(.tidyabm, values, is_characteristic = FALSE)
}
\arguments{
\item{.tidyabm}{a \code{tidyabm} object}

\item{values}{a named list where names reference charactistics/variables}

\item{is_characteristic}{if \code{FALSE} (the default), values are assumed to
belong to variables rather than characteristics}
}
\value{
a \code{tidyabm} object
}
\description{
Update an actual value from a characteristic or a variable
}
\keyword{internal}
