% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R, R/environment_grid.r
\name{add_agents}
\alias{add_agents}
\alias{add_agents.tidyabm_env}
\alias{add_agents.tidyabm_env_grid}
\title{Add agents to the environment}
\usage{
add_agents(.tidyabm, agent, n, ...)

\method{add_agents}{tidyabm_env}(.tidyabm, agent, n, ...)

\method{add_agents}{tidyabm_env_grid}(.tidyabm, agent, n, initial_position = NULL, ...)
}
\arguments{
\item{.tidyabm}{the \code{tidyabm_env} object to which agents should be added}

\item{agent}{an \code{tidyabm_agent} object (the blueprint) to replicate from}

\item{n}{number of agents in the style of \code{agent} to add}

\item{...}{other arguments to pass to sub functions}

\item{initial_position}{either \code{NULL} (the default) where agents are placed
randomly and independent from each other or a function with two parameters
\code{me} (the agent to be placed) and \code{abm} (the current environment with
previous agents already placed) that gets called once per agent to place
and which has to return a vector of length 2 that will be interpreted as
\code{x} and \code{y} coordinates (i.e., \code{c(x, y)}).
Here is an example to position agents in subsequent order via a function:
coordinate_counter <- c(0, 1)
create_grid_environment(seed = 892,
size = 3) \%>\%
add_agents(create_agent(),
n = 3,
initial_position = \(me, abm) {
coordinate_counter\link{1} <- coordinate_counter\link{1} + 1
if (coordinate_counter\link{1} > 3) {
coordinate_counter\link{1} <- 1
coordinate_counter\link{2} <- coordiante_counter\link{2} + 1
}
return(coordinate_counter)
})}
}
\value{
\code{tidyabm} object
}
\description{
Create an agent first (through \link{create_agent}). This agent can be seen as
a blueprint. Then, use this function here to add as many replicates of this
blueprint agent as necessary. They will be added independently so that they
do not interfere with each other throughout runtime.
}
\examples{
a1 <- create_agent() \%>\%
  set_characteristic(age = 'young')
a2 <- create_agent() \%>\%
  set_characteristic(age = 'old')
m <- create_grid_environment(seed = 1024, size = 50) \%>\%
  add_agents(a1, 100) \%>\%
  add_agents(a2, 100)

}
