% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{add_rule}
\alias{add_rule}
\alias{add_rule.tidyabm}
\title{Add a rule (i.e., an action that is performed under certain conditions)}
\usage{
add_rule(.tidyabm, .label, .consequence, ...)

\method{add_rule}{tidyabm}(.tidyabm, .label, .consequence, ...)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm} object}

\item{.label}{string describing the rule}

\item{.consequence}{function to be executed if all conditions apply. Use
a function with two arguments, \code{me} (which is the \link{tidyabm} object at that
specific point in time) and \code{abm} (which is the whole \link{tidyabm} object).
Note that for \link{tidyabm_env} objects these two arguments are the same but
for \link{tidyabm_agent} objects they are different in that the first (\code{me}) is
the current agent and the second (\code{abm}) is the whole environment model.
\strong{Importantly}, for any changes to apply the function has to return the
updated \code{me} (i.e., a \link{tidyabm_agent} object for agent rules and a
\link{tidyabm_env} object for environment rules). If the function misses to do
so, any changes will be ignored and a warning will be issued during
runtime. If you write your own functions or need to provide additional
arguments to functions, use the style of anonymized functions directly
in-line (through \verb{function(me, abm) ... return(me)} or
\verb{\\(me, abm) ... return(me)}).}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> expressions that
return a logical value and are defined in terms of all characteristics
and variables of \code{me} (which is the \link{tidyabm} object to which the rule
was applied to). If multiple expressions are included, they are combined
with the \code{&} operator.}
}
\value{
a \link{tidyabm} object
}
\description{
todo

For helping utilities and examples, see the full documentation,
particularly per environment type.
}
\examples{
create_agent() \%>\%
  set_characteristic(age = 15) \%>\%
  add_rule('check minority', age >= 18, \(me, abm) end(ebm))

create_grid_environment(seed = 1268, size = 5) \%>\%
  add_rule('no more agents', 'todo')

}
