% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm_class.R
\name{remove_rule}
\alias{remove_rule}
\alias{remove_rule.tidyabm}
\title{Remove a certain rule}
\usage{
remove_rule(.tidyabm, label)

\method{remove_rule}{tidyabm}(.tidyabm, label)
}
\arguments{
\item{.tidyabm}{a \link{tidyabm} object}

\item{label}{string used to describe the rule}
}
\value{
a \link{tidyabm} object
}
\description{
Remove a certain rule
}
\examples{
  add_rule('minors move',
           age >= 18,
           .consequence = \(me, abm) {
             spot <- grid_get_free_neighboring_spots(me, abm) \%>\%
               dplyr::slice_sample(n = 1)
             grid_move(me, abm,
                       new_x = spot$.x,
                       new_y = spot$.y) \%>\%
               return()
           })
  remove_rule('minors move')

create_grid_environment(seed = 1268, size = 5) \%>\%
  add_variable(n_agents = \(me, abm) = nrow(convert_agents_to_tibble(me))) \%>\%
  add_rule('no more agents',
           n_agents == 0,
           .consequence = stop_abm) \%>\%
  remove_rule('no more agents')

}
