% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{tick}
\alias{tick}
\alias{tick.tidyabm_env}
\title{Perform one single iteration}
\usage{
tick(.tidyabm, verbose = TRUE, visualize = FALSE, ...)

\method{tick}{tidyabm_env}(.tidyabm, verbose = TRUE, visualize = FALSE, ...)
}
\arguments{
\item{.tidyabm}{the \link{tidyabm_env} object}

\item{verbose}{if TRUE (the default), a status message is printed at the end}

\item{visualize}{if TRUE, each tick ends with a visualization if this has
been implemented for the respective environment; default is FALSE}

\item{...}{further arguments to pass on to visualize}
}
\value{
a \link{tidyabm_env} object
}
\description{
A single iterations does the following (in this order):
\enumerate{
\item randomize agent order
\item go through randomized agents and ...
a. re-calculate agent variables (in the order they were added)
b. check agent rules (in the order they were added)
\item re-calculate environment variables (in the order they were added)
\item check environment rules (in the order they were added)
}
}
\examples{
create_grid_environment(seed = 46444, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init() \%>\%
  tick()

}
\seealso{
\link{iterate}
}
