% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyabm-package.R
\docType{package}
\name{tidyabm-package}
\alias{tidyabm}
\alias{tidyabm-package}
\title{tidyabm: Set Up, Run, and Document Agent-Based Models (ABM) the Tidy Way}
\description{
tidyABM is designed to streamline the process of setting up, running, and documenting Agent-Based Models (ABM) using the principles of tidy data. It thus allows researchers to easily organize, manipulate, and visualize ABMs in a structured, coherent, and intuitive manner, built on tidyverse conventions to efficiently build and explore ABMs while ensuring readability, reproducibility, and transparency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datenfruehstueck/tidyabm}
  \item Report bugs at \url{https://github.com/datenfruehstueck/tidyabm/issues}
}

}
\author{
\strong{Maintainer}: Mario Haim \email{haim@ifkw.lmu.de} (\href{https://orcid.org/0000-0002-0643-2299}{ORCID})

}
\keyword{internal}
