% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R, R/environment_grid.r
\name{visualize}
\alias{visualize}
\alias{visualize.tidyabm_env_grid}
\title{Visualize an environment's current state}
\usage{
visualize(.tidyabm, ...)

\method{visualize}{tidyabm_env_grid}(.tidyabm, color = NULL, shape = NULL)
}
\arguments{
\item{.tidyabm}{the \link{tidyabm_env} object}

\item{...}{other arguments passed to particular types of environment}

\item{color}{specify an agent characteristic or variable by name to color
agents according to this specific characteristic/variable (default is not
to apply any colors)}

\item{shape}{specify an agent characteristic or variable by name to shape
agents according to this specific characteristic/variable (default is to
have all dots)}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
Visualization highly depends on the particular type of environment.
For a grid environment, for example, visualization resembles a rectangle
with agents placed in it.
The function returns a \link[ggplot2:ggplot]{ggplot2::ggplot} object, ready to be further
visualized/styled/manipulated.
}
\examples{
create_grid_environment(seed = 4583, size = 4) \%>\%
  add_agents(create_agent(), 2) \%>\%
  init() \%>\%
  tick() \%>\%
  visualize()

}
